Update code mapping priority

+++ src/main/java/net/arnx/jef4j/util/CharRecord.java
@@ -41,7 +41,7 @@ public class CharRecord implements Record, Serializable {
 		return (pattern & (char)(1 << (15 - pos))) != 0;
 	}
 	
-	public int get(int pos) {
+	public long get(int pos) {
 		if (pos == '\uFFFF') {
 			return array[pos];
 		}
