+++ src/main/java/net/arnx/jef4j/util/Record.java
@@ -2,5 +2,5 @@ package net.arnx.jef4j.util;
 
 public interface Record {
 	public boolean exists(int pos);
-	public int get(int pos);
+	public long get(int pos);
 }
